requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
python-multipart>=0.0.9
anthropic>=0.50
openpyxl>=3.1
biopython>=1.83
pydantic>=2.5
chromadb>=0.5
